app/requirements.txt
blinker==1.7.0
cffi==1.16.0
charset-normalizer==3.3.2
click==8.1.7
cryptography==42.0.5
Flask==3.0.3
itsdangerous==2.2.0
Jinja2==3.1.3
MarkupSafe==2.1.5
mysql-connector==2.2.9
pdfminer.six==20231228
pdfplumber==0.11.0
pillow==10.3.0
pycparser==2.22
pypdf==4.2.0
pypdfium2==4.29.0
typing_extensions==4.11.0
Werkzeug==3.0.2
